Fixed typos and mingw stuff.
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@735 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CSoftwareDriver2.cpp
@@ -1541,7 +1541,7 @@ void CSoftwareDriver2::lightVertex ( s4DVertex *dest, const S3DVertex *source )
 		}
 		else
 		{
-			pf = powf(dotHV, Material.org.Shininess );
+			pf = (f32)pow(dotHV, Material.org.Shininess );
 		}
 
 		// accumulate ambient
